probe command remove dependency on uuid package
